VACMS-15279: Utilize MenuFormAlter trait in va_gov_home
