+++ docroot/modules/custom/va_gov_home/src/EventSubscriber/FormEventSubscriber.php
@@ -4,6 +4,7 @@ namespace Drupal\va_gov_home\EventSubscriber;
 
 use Drupal\core_event_dispatcher\Event\Form\FormAlterEvent;
 use Drupal\core_event_dispatcher\FormHookEvents;
+use Drupal\va_gov_header_footer\Traits\MenuFormAlter;
 use Drupal\va_gov_user\Service\UserPermsService;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
@@ -11,6 +12,8 @@ use Symfony\Component\EventDispatcher\EventSubscriberInterface;
  * VA.gov home event subscriber.
  */
 class FormEventSubscriber implements EventSubscriberInterface {
+
+  use MenuFormAlter;
 
   /**
    * The VA user permission service.
@@ -27,6 +30,15 @@ class FormEventSubscriber implements EventSubscriberInterface {
    */
   public function __construct(UserPermsService $permsService) {
     $this->permsService = $permsService;
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public static function getSubscribedEvents(): array {
+    return [
+      FormHookEvents::FORM_ALTER => ['formAlter'],
+    ];
   }
 
   /**
@@ -56,79 +68,6 @@ class FormEventSubscriber implements EventSubscriberInterface {
       ->hubMenuHideExpanded($form)
       ->hubMenuHideViewMode($form, $admin)
       ->hubMenuHideParentLink($form, $admin);
-  }
-
-  /**
-   * Hides the attributes form field for home page hub list link form.
-   *
-   * @param array $form
-   *   The form element array.
-   *
-   * @return $this
-   */
-  public function hubMenuHideAddtributes(array &$form): static {
-    if (!empty($form['options']['attributes'])) {
-      $form['options']['attributes']['#access'] = FALSE;
-    }
-    return $this;
-  }
-
-  /**
-   * Hides the expanded form field for home page hub list link form.
-   *
-   * @param array $form
-   *   The form element array.
-   *
-   * @return $this
-   */
-  public function hubMenuHideExpanded(array &$form): static {
-    if (!empty($form['expanded'])) {
-      $form['expanded']['#access'] = FALSE;
-    }
-    return $this;
-  }
-
-  /**
-   * Hides the view_mode form field for home page hub list link form.
-   *
-   * @param array $form
-   *   The form element array.
-   * @param bool $admin
-   *   TRUE if current user is an administrator.
-   *
-   * @return $this
-   */
-  public function hubMenuHideViewMode(array &$form, bool $admin): static {
-    if (!empty($form['view_mode'])) {
-      $form['view_mode']['#access'] = $admin;
-    }
-    return $this;
-  }
-
-  /**
-   * Hides the parent link form field for home page hub list link form.
-   *
-   * @param array $form
-   *   The form element array.
-   * @param bool $admin
-   *   TRUE if current user is an administrator.
-   *
-   * @return $this
-   */
-  public function hubMenuHideParentLink(array &$form, bool $admin): static {
-    if (!empty($form['menu_parent'])) {
-      $form['menu_parent']['#access'] = $admin;
-    }
-    return $this;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public static function getSubscribedEvents(): array {
-    return [
-      FormHookEvents::FORM_ALTER => ['formAlter'],
-    ];
   }
 
 }
